In sky130hd fastroute.tcl use "set_macro_extension 1"
The LEF macros used in chameleon have the property that the m2 obs is
larger than then m1 obs.  This leaves a region where only planar routing
is possible as m2 is blocked, and m1 is the min routing layer.  To
avoid this the macro extension blocks that hard to route region.  It
is overly conservative but acceptable.  In the future we may have
grt automaticaly block those regions and then this could be reverted.

Signed-off-by: Matt Liberty <mliberty@eng.ucsd.edu>