Add OpenPOWER Microwatt core
Microwatt is a 64 bit OpenPOWER core written in VHDL. It includes an
IEEE 754 double-precision binary floating-point unit as well as
supervisor support that allows it to run Linux. There are hard macros
for the 2 multipliers (integer and floating point), the 2 cache RAMs
(icache and dcache) and the 4kB main RAM.

The VHDL has been converted to Verilog using GHDL and Yosys. At the
point at which OpenROAD gets native VHDL support we can switch to using
that.

Signed-off-by: Anton Blanchard <anton@linux.ibm.com>