github: add devcontainer configuration
the idea is that the community can collaborate on this
configuration such that all that is required to test and
develop OpenROAD-flow-scripts is to fire up a github codespaces
from any brows.

This is just a seed.

Further down the road, improvements to this config will
make it faster and easier to set up OpenROAD-flow-scripts.

Rough idea for use-case:

1. Click on "code" in github and create github codespace
2. Run: sudo tools/OpenROAD/etc/DependencyInstaller.sh
3. Run: build_openroad.sh --local
4. . setup_env.sh
5. To create a launch configuraiton in tools/OpenROAD/.vscode/launch.settings,
   run: make DESIGN_CONFIG=designs/asap7/aes/config.mk vscode_debug_pdn
4. Run: code tools/OpenROAD
5. hit f7 to build
6. Set breakpoint, press f5, wait for the debugger to hit the breakpoint for
   the failing stage in #5

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>