designs/asap7: mock very large array
This is a test of a large, 175000 um^2, but simple, design.

A simple macro is repeated in an array. All wires at
the top level should be simple horizontal or vertical routes.

PDN insertion for the top level design takes 432 seconds
on my machine, but with the hack below, 8 seconds.

https://github.com/The-OpenROAD-Project/OpenROAD/pull/2795

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>