Makefile: use ln -fs instead of cp to save space
this starts to matter for .odb files in the gigabytes size,
which will happen e.g. in exploration of mm^2 sized
ASAP7 designs with lots of empty space.

Empty space on a die can be common when exploring floorplans
and starting with an overgenerous placement of macros.

make works with softlinks: when the date is identical to
the target, i.e. it is up to date.

Files are only copied "forward" in the flow, so
this is safe.

tar works with hardlinks and softlinks: if the file does not exist in the
archive, then the full file is archived. Subsequent references
to the same file gets a hardlink. This should see "make issue"
sizes reduced.

A symlink is overwritten when a file is opened in write & truncate
mode, thus not changing the file that the symlink points to. This
is unlike a hard link.

Relative paths are used, such that the results folder
can be moved.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>