makefile: add vars target
This allows generating vars.sh/tcl/gdb files, which is useful
when debugging, openroad command line can be copied and pasted
after sourcing vars.sh.

Also, it could be useful in separating the concern of executing
ORFS stages and checking dependencies, opening the road implementing
Bazel support on top of ORFS with a minimal maintenance burden
for ORFS and the Bazel support.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>