makefile: yosys 0.13 or newer required
use yosys that ships with OpenROAD if yosys is too old.

It is highly likely than e.g. Ubuntu has a too old Yosys
installed, so this avoids a common pitfall & confusion when relying
on ORFS to find the tools in the path, i.e. when not
using ". env.sh"

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>