global route: add policy on iteration reporting on hard runs
The default behavior if the user didn't specify GLOBAL_ROUTE_ARGS is to
first make an attempt with a few iterations, if this is a hard global
route, start from scratch, try harder and logging. The goal here is to make
ORFS handle the simple cases without fuzz and to make the harder
cases more easily debuggable without having to go spelunking in Tcl.

If GLOBAL_ROUTE_ARGS is specified, then we do only what the
GLOBAL_ROUTE_ARGS specifies.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>