fix a bug in asap7/riscv32i that has memory signals dangling (+ metrics)
In dmem.v the inter_dmem* signals weren't declared which meant they were
treated as single-bit signals instead of 32-bit busses!  All but one
bit of the ram outputs were dangling.

Fixing this makes the metrics much worse but that will have to be
tackled separately as the current result is clearly wrong.

Signed-off-by: Matt Liberty <mliberty@precisioninno.com>