makefile: add OPENROAD_GUI_ARGS
the usecase is to run openroad in console mode.

This can be used by humans, but also for scripting.

For instance: inject a command or two and then parse the
output. Convenient for quick and dirty Python scripts to plot something,
for instance.

To start OpenROAD in console mode with the final result:

make OPENROAD_GUI_ARGS="" gui_final

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>