makefile: add do-gds
separate the concerns of checking dependencies and
making the gds file.

Example use-case: make a wafer thin Bazel wrapper on top of
ORFS to get artifact support for macros that can be in the
hundreds of megabytes range.

WRAP_LEFS are not supported for do-gds as this is only used
by gf12 and it is under NDA. Also, WRAP_LEFS should be
retired and not invested further in.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>