makefile: improve logging, add stage label for Running line
the stage filename stems and which outputs are produced
is somewhat idiosyncratic and in some places inconsistent
and ambigious(such as finish/final).

Therefore simply log "Running pdn.tcl, stage 2_floorplan_pdn".

This is accurate and precise as far as we can be and remains
accurate if filename stems and inputs/outputs are made
more consistent in the future.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>