synthesis: allow blackboxing macros
If a macro is specified in ADDITIONAL_LEFS/LIBS and it is also present
in the Verilog files, use MACROS to list those macros so as to have the
module be marked as blackbox.

A Verilog file can contain more than one module, so it isn't always
convenient to exclude a .v/sv file to have a module marked as
black box.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>