makefile: delete preview macro stuff
I was excited about this feature, then I ended up never using it and
forgot all about it.

Perhaps someone is using it, but the only way to know is to delete it
and see if anyone protests in pull request or after it is merged,
in which case it is easily restored or reimplemented.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>