synth: add canonicalization step before synthesis
This saves hours on larger designs when unecessary macro rebuilds
are avoided and makes the build scripts less brittle.

this will clean out unused modules and produce the a canonical
$(RESULTS_DIR)/1_synth.json file.

This file will not change hash when unrelated changes to the design changes,
such as when non-synthesis(simulation) files are mixed into VERILOG_FILES.

This is particularly useful when doing synthesis on macros out of a folder of
Verilog files and you don't want to carefully groom the list of VERILOG_FILES.

This also allows bazel-orfs to handle actual dependencies for a folder
of .v/.sv files in much the same fashion that CMake can figure out
dependencies given a list of .cpp and header files.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>