global route: now consistent with detailed routing w.r.t. congestion failures
Just like detailed routing, a global route that fails with congestion, is not
a build failure(as in exit code non-zero), it is a successful(as in zero exit code)
global route that produce reports detailing the problem.

Detailed route will not proceed, if there is global routing congestion

This allows build systems, such as bazel, to create artifacts for global
and detailed route, even if the operation had problems, without having
know about the semantics between global and detailed route.

Considering that global and detailed route can run for a long time and
use a lot of memory, this allows inspecting results on a laptop for
a build that ran on a server.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>