makefile: add do-yosys-flatten-policy stage
This replaces iffy code in the Makefile. The flattening policy is a script
that is created after canonicalization and prior to synthesis. It is
sourced from synthesis to mark modules not to be flattened.

The Makefile is now more idiomatic: input files to a target and output files,
no iffy code.

This also works better with Bazel and paves the way for storing
the flattening policy in git, shaving hours of build times.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>