mock-cpu: remove set_input/output_delays
Not idiomatic for the use-case where the mock-cpu
would be connected into an SoC where timing closure
is checked.

setting set_input_delay in particular is problematic
because it can lead to hold cells which should not be
used at all, or if they have to be used, they should
be inserted where the mock-cpu is used at the SoC level,
not inside the mock-cpu macro.

set_input/output_delay would be more approperiate at
the top level of an SoC on the pins connecting to
the outside world.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>