synthesis: better reports for exploring MAX_UNGROUP_SIZE
Seperate the concern of collecting area of modules and implementing
a policy.

First a report is collected, such as:

$ cat results/asap7/riscv32i/base/keep_hierarchy.tcl
0.291600 $paramod\mux2\WIDTH=s32'00000000000000000000000000000001
0.306180 \magcompare2c
0.641520 \magcompare2b
0.801900 $paramod\mux2\WIDTH=s32'00000000000000000000000000000100
0.991440 $paramod\mux2\WIDTH=s32'00000000000000000000000000000101
2.084940 \ROM
3.076380 \aludec
4.009500 \maindec
6.794280 $paramod\mux2\WIDTH=s32'00000000000000000000000000100000
6.794280 $paramod\mux2\WIDTH=s32'00000000000000000000000000100000

In synthesis, MAX_UNGROUP_SIZE is used to select which
of these modules to keep and which to flatten.

For some design there is little guesswork involved when looking
at this report and picking a MAX_UNGROUP_SIZE.

The names of targets in the makefile, variables and report
names could be revised to reflect this seperation of concerns
to improve readability, but leave that for a followup PR.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>