cleanup: introduce place_density_with_lb_addon proc
this replaces unique pattern where a file is sourced and it
acts more or less as a proc, setting a variable.

Made logging a bit more consistent, added it for global_placement
invocation in global_place_skip_io. This makes it evident in the
code that place_density_with_lb_addon is logged as soon as
it is calculated.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>