open: fix non-aborting incorrect operation when there is no routing
When using open.tcl from automation, if there is no routing available
as of stage 5 and on, fail rather than log and continue.

The user experience when running 'make gui_grt' is that an error
is shown and a console prompt. The user can type 'gui::show' to
show the GUI.

Ideally, 'make gui_grt' would run with the GUI hidden and show
the GUI upon first error or warning, but that requires
an update to OpenROAD.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>