asap7/ethmac_lvt: constrain pins to edges to make test case less volatile
slightly different initial conditions can lead to
completely different pin placements, which can
lead to real global routing failures.

Instead of letting pins be on any edge, constrain the
test case a bit by having wishbone (a guess as to what
wb_ means) to left edge and the rest of the pins to
the top edge.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>