mock-array: delete more harm than good README.md
Some of the most interesting documentation has been moved
to platforms/asap7/constraints.sdc

Regenerating the Verilog or changing parameters is there for
reference, it isn't really an end user feature. The Verilog
code commited to git is the only one that is actively studied
and run through regression tests.

Any change to the Verilog code would need to come with
careful updates to the config.mk for Element and mock-array.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>