variables: break out variables.mk file
The idea is to separate the ORFS variable support and make
variables and ORFS agnostic to build systems, even if
make is used to evaluate variables.

ORFS contains an implementation of a flow using make,
bazel-orfs implements a flow using Bazel.

Sometimes make is preferable, keep things simple, but for more advanced
flows, the simplicity of make introduces more problems than it solves

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>