open_synth: make it work with OpenSTA
Modify scripts so that make open_synth can work with OpenSTA, this simplifies reporting issues. Turnaround times on real test cases is long, so it is nice to have this debugged in ORFS.

To examine timing of a design using OpenSTA:

    make OPENROAD_EXE=../tools/install/OpenROAD/bin/sta open_synth

To create an OpenSTA standalone issue, run:

    make open_issue

Tinker with run-*.sh script, create a small script, source scripts/sta-synth.tcl and create test-case.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>