generate-docs: stable order
an ordering was dependent on set of dictionary keys
from yaml parsing of a dictionary.

We want a stable ordering and that is a lot of
implementation detail to base that stable ordering
feature on, so sort to make it clear in the code
that we have a stable order.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>