Obtain bash path through /usr/bin/env in flow scripts.
Unfortunately, not all systems have bash in /bin. One notable example is nixos.
When entering a development shell in nixos, the current version of the scripts
will fail complaining that "something is wrong at line 1". The issue is that
there is no /bin/bash on NixOS or more generally on the system.

Signed-off-by: Marco Minutoli <marco.minutoli@pnnl.gov>