make: check if binaries are valid before continuing
The current logic in variables.mk results in OPENROAD_EXE and YOSYS_EXE
to always be set to something, either the absolute path to tools/install
or the value from the parent env. With this behavior, when calling make
versions.txt the if statements are always true and the "not available"
is never printed to the versions.txt file. If the bins are not valid it
just prints an empty string.

Before this change, calling make versions.txt would fail silently ($? == 0)
and other make commands would fail when trying to run openroad or yosys.

After this change, calling any make command without having a valid
openroad or yosys binary will fail with an error message, which is the
same behaviour already implemented for klayout.

Signed-off-by: Vitor Bandeira <vvbandeira@precisioninno.com>